+++ CHANGELOG.md
@@ -7,9 +7,9 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Bug Fixes
 
+- Fix ensure QueuingTransactionReceiptProcessor executor is properly shut down to avoid JVM hang [#2265](https://github.com/LFDT-web3j/web3j/pull/2265)
 - Fix Transaction serialization failure when fee fields are null [#2293](https://github.com/LFDT-web3j/web3j/pull/2293)
 - Fix NullPointerException when deriving child keys from public-only Bip32ECKeyPair [#2284](https://github.com/LFDT-web3j/web3j/pull/2284)
-
 
 ### Features